Defaulting Asset handle to file name.
Sick of typing in a handle that's identical to the file name.
Do you need handle-less assets? If so, this isn't for you.
Or maybe you can just remove the handle later, since it just defaults on create.

Also, making sure we don't override a supplied handle. That would be dumb.