Allowing Asset attached data to possibly be in s3
Now you can choose between filesystem (with default values) and
s3 (with a hard-coded path and credentials location -- deal with it).
You just provide the bucket name and credentials, and the model
switches over.

The specs are a bit nasty because this is all set once the model class
is loaded, so they have to be reset in this heavy-handed way to ensure
the settings take.