Adding asset_link view helper.
This takes a asset handle and returns the appropriate show link, either
the s3 one for s3 storage or the simple /assets/HANDLE for file storage.

I wish I could call this asset_path, but the routing helper gets in the
way. So it goes.

And speaking Rails getting in the way, apparently I need a helper at the
app level to get this loaded, hence the change to the plugin-install
template.